Platform: made the platform autoparent to the play area transform
According to the VRTK docs, moving around in the simulator setup moves both the play area and the headset, so in the simulator it looks like the platform is parented to the camera. Hopefully that's not the case in the actual VR setup, but I can only test that tomorrow on the actual equipment.